[MNG-7250] Upgrade Sisu Inject/Plexus to 0.3.5
Since we have replaced the old JSR 250 library with javax.annotation library we
go straight to 1.3.2 because Maven Resolver 1.7.0 introduced this already as a
transitive dependency in f749be3a6e452579fb2df155ea7a82f266810ecd.
The only difference between version 1.2 and 1.3.2 is a Java upgrade from 6 to 8.

This closes #539